cortex 8 's windows support
my file is in windows format,It make git think all file has changed....

+++ src/IECore/ContrastSmoothSkinningWeightsOp.cpp
@@ -207,7 +207,7 @@ void ContrastSmoothSkinningWeightsOp::modify( Object * object, const CompoundObj
 		locks.resize( skinningData->influenceNames()->readable().size(), false );
 	}
 	
-	std::vector<int64_t> vertexIds;
+	std::vector<int> vertexIds;
 	m_vertexIdsParameter->getFrameListValue()->asList( vertexIds );
 	
 	// make sure all vertex ids are valid
