+++ src/IECore/LimitSmoothSkinningInfluencesOp.cpp
@@ -223,7 +223,7 @@ void LimitSmoothSkinningInfluencesOp::modify( Object * object, const CompoundObj
 	// Zero specific influences for all points
 	else if ( mode == LimitSmoothSkinningInfluencesOp::Indexed )
 	{
-		std::vector<int64_t> indicesToLimit;
+		std::vector<int> indicesToLimit;
 		m_influenceIndicesParameter->getFrameListValue()->asList( indicesToLimit );
 		std::vector<bool> limitIndex( skinningData->influenceNames()->readable().size(), false );
 		for ( unsigned i=0; i < indicesToLimit.size(); i++ )
